First crack at extension functions.
Everything compiles and there are some complicated diamond dependencies going on. It would be way easier if we had composition instead of inheritance. As-is KotlinExtensionFunctionWithReturn has to choose whether it inherits its implementation from KotlinExtensionFunction or KotlinFunctionWithReturn. This case was pretty even, but in general the answer will be to either pick the path with the least code duplication, or pull the entire implementation into a single master class that can handle all of it and implements all of the interfaces.