First pass at if blocks
This was a big change, but it looks like if blocks are pretty solid. You can use them in a continuation style (e.g. `.ifBlock(condition).doReturn(...).elseBlock().doReturn(...)`), or you can use `define` in Kotlin to be able to pass longer blocks. The `define` form automatically advances to an IfContinuation once you've closed the block, making a further `define` call unnecessary.